Extract CaptureEnv type into a submodule
